checkout done

+++ cart.php
@@ -1,6 +1,10 @@
 <?php
-    include 'config.php';
+    require_once 'cekLogin.php';
     $user_id = $_SESSION['user']['id'];
+    
+    if($_SESSION['user']['address']==null){
+        header('location: '.$base_url.'/myAccount.php');
+    }
 ?>
 
 <!DOCTYPE html>
@@ -49,48 +53,56 @@
                             $sql = "SELECT CART.*,CART.id as 'cart_id',PRODUCT.* FROM CART JOIN PRODUCT on PRODUCT.id = CART.product_id WHERE user_id = $user_id";
                             $query = mysqli_query($con,$sql);
                             $grandtotal = 0;
-                            while($re = mysqli_fetch_assoc($query)) {
-                                $image = explode(",",$re['image'])[0];
-                                $name = $re['name'];
-                                $id = $re['cart_id'];
-                                $variance = $re['variance'];
-                                $price = $re['subtotal'];
-                                $quantity = $re['quantity'];
-                                $grandtotal += $re['subtotal'];
-                                $stocks = $re['stocks'];
-                            ?>
-                            <div class="col-md-12 row my-3">
-                                <div class="col-md-3">
-                                    <div class="card" style="width: 8rem;height:fit-content;">
-                                        <img src="<?=$base_url;?>/assets/img/<?=$image;?>" class="card-img-top cart-img" style="max-height: 100px;width:auto">
+                            $n = mysqli_num_rows($query);
+
+                            if($n>0) {
+                                while($re = mysqli_fetch_assoc($query)) {
+                                    $image = explode(",",$re['image'])[0];
+                                    $name = $re['name'];
+                                    $id = $re['cart_id'];
+                                    $variance = $re['variance'];
+                                    $price = $re['subtotal'];
+                                    $quantity = $re['quantity'];
+                                    $grandtotal += $re['subtotal'];
+                                    $stocks = $re['stocks'];
+                                ?>
+                                <div class="col-md-12 row my-3">
+                                    <div class="col-md-3">
+                                        <div class="card" style="width: 8rem;height:fit-content;">
+                                            <img src="<?=$base_url;?>/assets/img/<?=$image;?>" class="card-img-top cart-img" style="max-height: 100px;width:auto">
+                                        </div>
                                     </div>
-                                </div>
-                                <div class="col-md-7 text-white">
-                                    <h5><?=$name;?></h5>
-                                    <div class="d-flex justify-content-between">
-                                    Size : <?=$variance;?><br>
-                                    <span class="text-success"> Rp.<?=number_format($price);?></span>
+                                    <div class="col-md-7 text-white">
+                                        <h5><?=$name;?></h5>
+                                        <div class="d-flex justify-content-between">
+                                        Size : <?=$variance;?><br>
+                                        <span class="text-success"> Rp.<?=number_format($price);?></span>
+                                        
+                                        </div>
+                                    <div class="d-flex mt-2 justify-content-between">
+                                            Quantity : <?=$quantity;?>
+                                            <a href="<?=$base_url;?>/proses/deleteCart.php?id=<?=$id;?>"><button class="btn btn-sm btn-danger"><i class="fas fa-trash"></i></button></a>
+                                        </div>
+                                    </div>
                                     
-                                    </div>
-                                   <div class="d-flex mt-2 justify-content-between">
-                                        Quantity : <?=$quantity;?>
-                                        <a href="<?=$base_url;?>/proses/deleteCart.php?id=<?=$id;?>"><button class="btn btn-sm btn-danger"><i class="fas fa-trash"></i></button></a>
-                                    </div>
                                 </div>
-                                
-                            </div>
-                        <?php
+                                <?php
                             }
-                        ?>
+                                    } else {
+                                        echo '<h3 class="text-white text-center mb-5">..Nothing here, lets shoping..';
+                                    }
+                                ?>
                     </div>
                 </div>
             </div>
             <div class="col-md-4" style="min-height: 100vh;">
                 <div class="card bg-dark">
                     <div class="card-body text-white">
-                        <h6 class="mt-2">Payemnt</h6>
-                        <hr class="bg-light mb-5">
-                        <form action="<?=$base_url;?>/proses/cekCoupun.php" method="post">
+                        <h6 class="mt-2">Payemnt Info</h6>
+                        <?=$_SESSION['user']['name'];?> &nbsp; &nbsp;<?=$_SESSION['user']['phone_number'];?>
+                            <p><?=$_SESSION['user']['address'];?></p>
+                        <hr class="bg-light">
+                        <form action="<?=$base_url;?>/proses/cekCoupun.php" method="post" class="mt-5">
                             <div class="row container my-5">
                                 <div class="col-md-9">
                                     <input type="text" name="coupun" placeholder="coupun.." id="coupun" class="form-control">
@@ -106,11 +118,11 @@
                         <div class="container d-flex justify-content-between">
                             <div>
                                 Subtotal &nbsp; :<br>
-                                <span class="discont <?php if(isset($_SESSION['disc'])) echo 'd-block'; else echo 'd-none';?>d-none">Discount &nbsp;:</span>
+                                <span class="discont <?php if(isset($_SESSION['disc'])) echo 'd-block'; else echo 'd-none';?>">Discount &nbsp;:</span>
                             </div>
                             <div>
                                 Rp. <?=number_format($grandtotal);?><br>
-                                <span class="discont <?php if(isset($_SESSION['disc'])) echo 'd-block'; else echo 'd-none';?>d-none"><?php if(isset($_SESSION['disc'])) echo'Rp. '. number_format($_SESSION['disc'])?></span>
+                                <span class="discont <?php if(isset($_SESSION['disc'])) echo 'd-block'; else echo 'd-none';?>"><?php if(isset($_SESSION['disc'])) echo'Rp. '. number_format($_SESSION['disc'])?></span>
                             </div>
                         </div>
                         <hr class="bg-light my-5">
@@ -126,7 +138,10 @@
                         </div>
                         <div class="container">
                             <div class="col-md-12">
-                                <button class="btn-block btn btn-secondary">Checkout</button>
+                                <form action="<?=$base_url;?>/checkout.php" method="post">
+                                    <input type="hidden" name="grandtotal" value="<?=$grandtotal;?>">
+                                    <button type="submit" name="btn" class="btn-block btn btn-secondary">Checkout</button>
+                                </form> 
                             </div>
                         </div>
                     </div>
